commonize format function
