+++ lib/holiday_jp.js
@@ -1,12 +1,6 @@
 var package_info = require('./../package.json');
 var holidays = require('./holidays');
-
-function format(date) {
-  var year = date.getFullYear();
-  var month = ('0' + (date.getMonth() + 1)).slice(-2);
-  var day = ('0' + (date.getDate())).slice(-2);
-  return (year + '-' + month + '-' + day);
-}
+var format = require('./format');
 
 var holiday_jp = {
   VERSION: package_info.version,
